+++ packages/manager/apps/container/src/container/nav-reshuffle/sidebar/navigation-tree/node.ts
@@ -27,6 +27,7 @@ export type Node = {
   hideIfEmpty?: boolean;
   badge?: string;
   forceVisibility?: boolean;
+  forceNavigate?: boolean;
   url?: string | Record<string, string>;
   isExternal?: boolean;
   hidden?: boolean;