+++ packages/manager/apps/key-management-service/src/components/layout-helpers/Dashboard/GeneralInformationsTiles/InformationsTile.constants.ts
@@ -1,3 +1,0 @@
-export const KMIP_LABEL = 'KMIP';
-export const KMIP_RSA_LABEL = `${KMIP_LABEL} (RSA)`;
-export const SWAGGER_UI_LABEL = 'Swagger UI';