+++ packages/manager/apps/zimbra/src/pages/dashboard/MailingLists/MailingListSettings.page.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useNotifications } from '@ovh-ux/manager-react-components';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useSearchParams } from 'react-router-dom';
@@ -20,6 +20,7 @@ import {
   ODS_INPUT_TYPE,
   ODS_MESSAGE_COLOR,
   ODS_TEXT_PRESET,
+  OdsSelectChangeEvent,
 } from '@ovhcloud/ods-components';
 import { ApiError } from '@ovh-ux/manager-core-api';
 import { useMutation } from '@tanstack/react-query';
@@ -29,6 +30,8 @@ import {
   PageType,
   useOvhTracking,
 } from '@ovh-ux/manager-react-shell-client';
+import { Controller, SubmitHandler, useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
 import { useGenerateUrl, usePlatform } from '@/hooks';
 import {
   postZimbraPlatformMailingList,
@@ -40,7 +43,7 @@ import {
   getZimbraPlatformMailingListsQueryKey,
 } from '@/api/mailinglist';
 import { DomainType } from '@/api/domain';
-import { ACCOUNT_REGEX, OWNER_REGEX } from '@/utils';
+import { mailingListSchema, MailingListSchema } from '@/utils';
 import queryClient from '@/queryClient';
 import {
   ADD_MAILING_LIST,
@@ -48,7 +51,6 @@ import {
   CONFIRM,
   EDIT_MAILING_LIST,
 } from '@/tracking.constant';
-import { FormTypeInterface, useForm } from '@/hooks/useForm';
 
 const replyToChoices = [
   {
@@ -81,13 +83,13 @@ const moderationChoices = [
 ];
 
 // fetch this from api ?
-const languageList = ['FR', 'EN', 'ES'];
+export const languages = ['FR', 'EN', 'ES'] as const;
 
 export default function MailingListSettings({
-  domainList = [],
+  domains = [],
   editMailingListDetail = null,
 }: Readonly<{
-  domainList: DomainType[];
+  domains: DomainType[];
   editMailingListDetail: MailingListType;
 }>) {
   const { trackClick, trackPage } = useOvhTracking();
@@ -97,7 +99,7 @@ export default function MailingListSettings({
   const { platformId } = usePlatform();
   const [searchParams] = useSearchParams();
   const editMailingListId = searchParams.get('editMailingListId');
-  const organizationIdParam = searchParams.get('organizationId');
+  const organizationId = searchParams.get('organizationId');
   const trackingName = editMailingListId ? EDIT_MAILING_LIST : ADD_MAILING_LIST;
   const [selectedDomainOrganization, setSelectedDomainOrganization] = useState(
     '',
@@ -106,94 +108,10 @@ export default function MailingListSettings({
 
   const goBack = () => navigate(goBackUrl);
 
-  const { form, setForm, isFormValid, setValue } = useForm({
-    ...{
-      account: {
-        value: '',
-        defaultValue: '',
-        required: true,
-        validate: ACCOUNT_REGEX,
-      },
-      domain: {
-        value: '',
-        defaultValue: '',
-        required: true,
-      },
-      defaultReplyTo: {
-        value: '',
-        defaultValue: '',
-        required: true,
-      },
-      owner: {
-        value: '',
-        defaultValue: '',
-        required: true,
-        validate: OWNER_REGEX,
-      },
-      language: {
-        value: '',
-        defaultValue: '',
-        required: true,
-      },
-      moderationOption: {
-        value: '',
-        defaultValue: '',
-      },
-      subscriberModeration: {
-        value: '',
-        defaultValue: '',
-      },
-    },
-  });
-
-  const getDataBody = useCallback(
-    (formRef: FormTypeInterface) => {
-      const {
-        account: { value: account },
-        domain: { value: domain },
-      } = form;
-
-      let dataBody = {
-        email: `${account}@${domain}`,
-      };
-
-      Object.entries(formRef).forEach(([key, { value }]) => {
-        if (!['account', 'domain'].includes(key)) {
-          dataBody = { ...dataBody, [key]: value };
-        }
-      });
-
-      return dataBody;
-    },
-    [form],
-  );
-
-  useEffect(() => {
-    if (editMailingListDetail) {
-      const newForm: FormTypeInterface = form;
-      const {
-        email,
-        defaultReplyTo,
-        language,
-        owner,
-        moderationOption,
-      } = editMailingListDetail.currentState;
-      const [account, domain] = email.split('@');
-      newForm.account.value = account;
-      newForm.domain.value = domain;
-      newForm.defaultReplyTo.value = defaultReplyTo;
-      newForm.owner.value = owner;
-      newForm.language.value = language;
-      newForm.moderationOption.value = moderationOption;
-      setForm((oldForm) => ({ ...oldForm, ...newForm }));
-    }
-  }, [editMailingListDetail]);
-
-  const handleDomainChange = (selectedDomain: string) => {
-    const organizationLabel = domainList.find(
-      ({ currentState }) => currentState.name === selectedDomain,
+  const setSelectedOrganization = (e: OdsSelectChangeEvent) => {
+    const organizationLabel = domains.find(
+      ({ currentState }) => currentState.name === e?.detail?.value,
     )?.currentState.organizationLabel;
-    setValue('domain', selectedDomain);
     setSelectedDomainOrganization(organizationLabel);
   };
 
@@ -246,14 +164,47 @@ export default function MailingListSettings({
     },
   });
 
-  const handleSavelick = () => {
+  const {
+    control,
+    handleSubmit,
+    formState: { isDirty, isValid, errors },
+  } = useForm({
+    defaultValues: {
+      account: editMailingListDetail?.currentState?.email?.split('@')[0] || '',
+      domain: editMailingListDetail?.currentState?.email?.split('@')[1] || '',
+      defaultReplyTo: editMailingListDetail?.currentState?.defaultReplyTo || '',
+      owner: editMailingListDetail?.currentState?.owner || '',
+      language: editMailingListDetail?.currentState?.language || '',
+      moderationOption:
+        editMailingListDetail?.currentState?.moderationOption || '',
+      subscriberModeration:
+        editMailingListDetail?.currentState?.subscriberModeration || false,
+    },
+    mode: 'onTouched',
+    resolver: zodResolver(mailingListSchema),
+  });
+
+  const handleSavelick: SubmitHandler<MailingListSchema> = (data) => {
     trackClick({
       location: PageLocation.page,
       buttonType: ButtonType.button,
       actionType: 'action',
       actions: [trackingName, CONFIRM],
     });
-    addOrEditMailingList(getDataBody(form));
+
+    const { account, domain } = data;
+
+    const payload: Record<string, unknown> = {
+      email: `${account}@${domain}`,
+    };
+
+    Object.entries(data).forEach(([key, value]) => {
+      if (!['account', 'domain'].includes(key)) {
+        payload[key] = value;
+      }
+    });
+
+    addOrEditMailingList(payload);
   };
 
   const handleCancelClick = () => {
@@ -267,7 +218,10 @@ export default function MailingListSettings({
   };
 
   return (
-    <div className="w-full md:w-3/4 flex flex-col space-y-5">
+    <form
+      onSubmit={handleSubmit(handleSavelick)}
+      className="w-full md:w-3/4 flex flex-col space-y-5"
+    >
       <OdsText preset={ODS_TEXT_PRESET.paragraph}>
         {!editMailingListId
           ? t('zimbra_mailinglist_add_header')
@@ -276,57 +230,69 @@ export default function MailingListSettings({
       <OdsText preset={ODS_TEXT_PRESET.caption}>
         {t('common:form_mandatory_fields')}
       </OdsText>
-      <OdsFormField>
-        <label slot="label">
-          {t('zimbra_mailinglist_add_input_email_label')} *
-        </label>
-        <div className="flex">
-          <OdsInput
-            type={ODS_INPUT_TYPE.text}
-            name="account"
-            placeholder={t('zimbra_mailinglist_add_input_email_placeholder')}
-            hasError={form.account.hasError}
-            value={form.account.value}
-            defaultValue={form.account.defaultValue}
-            isRequired={form.account.required}
-            className="w-1/2"
-            data-testid="input-account"
-            onOdsBlur={({ target: { name, value } }) =>
-              setValue(name, value.toString(), true)
-            }
-            onOdsChange={({ detail: { name, value } }) => {
-              setValue(name, String(value));
-            }}
-          ></OdsInput>
-          <OdsInput
-            type={ODS_INPUT_TYPE.text}
-            name={'@'}
-            value={'@'}
-            isReadonly
-            isDisabled
-            className="input-at w-10"
-          ></OdsInput>
-          <OdsSelect
-            name="domain"
-            value={form.domain.value}
-            defaultValue={form.domain.defaultValue}
-            className="w-1/2"
-            hasError={form.domain.hasError}
-            isRequired={form.domain.required}
-            onOdsChange={(e) => handleDomainChange(e.detail.value)}
-            placeholder={t('common:select_domain')}
-            data-testid="select-domain"
-          >
-            {domainList?.map(({ currentState: domain }) => (
-              <option key={domain.name} value={domain.name}>
-                {domain.name}
-              </option>
-            ))}
-          </OdsSelect>
-        </div>
-      </OdsFormField>
-      {selectedDomainOrganization && !organizationIdParam && (
-        <OdsMessage color={ODS_MESSAGE_COLOR.information}>
+      <Controller
+        control={control}
+        name="account"
+        render={({ field: { name, value, onChange, onBlur } }) => (
+          <OdsFormField className="w-full" error={errors?.[name]?.message}>
+            <label htmlFor={name} slot="label">
+              {t('zimbra_mailinglist_add_input_email_label')} *
+            </label>
+            <div className="flex">
+              <OdsInput
+                type={ODS_INPUT_TYPE.text}
+                placeholder={t(
+                  'zimbra_mailinglist_add_input_email_placeholder',
+                )}
+                data-testid="input-account"
+                className="flex-1"
+                id={name}
+                name={name}
+                hasError={!!errors[name]}
+                value={value}
+                onOdsBlur={onBlur}
+                onOdsChange={onChange}
+              />
+              <OdsInput
+                type={ODS_INPUT_TYPE.text}
+                name="@"
+                value="@"
+                isReadonly
+                isDisabled
+                className="input-at w-10"
+              />
+              <Controller
+                control={control}
+                name="domain"
+                render={({ field }) => (
+                  <OdsSelect
+                    id={name}
+                    name={field.name}
+                    hasError={!!errors[field.name]}
+                    value={field.value}
+                    placeholder={t('common:select_domain')}
+                    onOdsChange={(e) => {
+                      field.onChange(e);
+                      setSelectedOrganization(e);
+                    }}
+                    onOdsBlur={field.onBlur}
+                    data-testid="select-domain"
+                    className="flex-1"
+                  >
+                    {domains?.map(({ currentState: domain }) => (
+                      <option key={domain.name} value={domain.name}>
+                        {domain.name}
+                      </option>
+                    ))}
+                  </OdsSelect>
+                )}
+              />
+            </div>
+          </OdsFormField>
+        )}
+      />
+      {selectedDomainOrganization && !organizationId && (
+        <OdsMessage color={ODS_MESSAGE_COLOR.information} isDismissible={false}>
           <OdsText preset={ODS_TEXT_PRESET.paragraph}>
             {t('zimbra_mailinglist_add_message_organization', {
               organization: selectedDomainOrganization,
@@ -334,145 +300,172 @@ export default function MailingListSettings({
           </OdsText>
         </OdsMessage>
       )}
-      <OdsFormField>
-        <label slot="label">{t('common:owner')} *</label>
-        <div className="flex">
-          <OdsInput
-            type={ODS_INPUT_TYPE.text}
-            name="owner"
-            placeholder={t('common:owner')}
-            hasError={form.owner.hasError}
-            value={form.owner.value}
-            defaultValue={form.owner.defaultValue}
-            isRequired={form.owner.required}
-            className="w-1/2"
-            data-testid="input-owner"
-            onOdsBlur={({ target: { name, value } }) =>
-              setValue(name, value.toString(), true)
-            }
-            onOdsChange={({ detail: { name, value } }) => {
-              setValue(name, String(value));
-            }}
-          ></OdsInput>
-        </div>
-      </OdsFormField>
-      <OdsFormField>
-        <label slot="label">
-          {t('zimbra_mailinglist_add_reply_to_label')} *
-        </label>
-        <div className="flex flex-col gap-4">
-          {replyToChoices.map(({ value, key }) => (
-            <div key={value} className="flex leading-none gap-4">
-              <OdsRadio
-                id={value}
-                name="defaultReplyTo"
-                value={value}
-                onOdsChange={(event) =>
-                  setValue('defaultReplyTo', event.detail.value)
-                }
-                data-testid={`radio-reply-to-${value}`}
-              ></OdsRadio>
-              <label htmlFor={value} className="flex flex-col">
-                <OdsText preset={ODS_TEXT_PRESET.paragraph}>{t(key)}</OdsText>
-              </label>
+      <Controller
+        control={control}
+        name="owner"
+        render={({ field: { name, value, onChange, onBlur } }) => (
+          <OdsFormField error={errors?.[name]?.message}>
+            <label htmlFor={name} slot="label">
+              {t('common:owner')} *
+            </label>
+            <OdsInput
+              placeholder={t('common:owner')}
+              className="w-1/2"
+              data-testid="input-owner"
+              type={ODS_INPUT_TYPE.text}
+              id={name}
+              name={name}
+              hasError={!!errors[name]}
+              value={value}
+              onOdsBlur={onBlur}
+              onOdsChange={onChange}
+            />
+          </OdsFormField>
+        )}
+      />
+      <Controller
+        control={control}
+        name="defaultReplyTo"
+        render={({ field: { name, value, onChange } }) => (
+          <OdsFormField error={errors?.[name]?.message}>
+            <label htmlFor={name} slot="label">
+              {t('zimbra_mailinglist_add_reply_to_label')} *
+            </label>
+            <div className="flex flex-col gap-4">
+              {replyToChoices.map((choice) => (
+                <div key={choice.value} className="flex leading-none gap-4">
+                  <OdsRadio
+                    id={choice.value}
+                    name={choice.value}
+                    value={value}
+                    isChecked={value === choice.value}
+                    onClick={() => onChange(choice.value)}
+                    data-testid={`radio-reply-to-${choice.value}`}
+                  ></OdsRadio>
+                  <label
+                    htmlFor={choice.value}
+                    className="flex flex-col cursor-pointer"
+                  >
+                    <OdsText preset={ODS_TEXT_PRESET.paragraph}>
+                      {t(choice.key)}
+                    </OdsText>
+                  </label>
+                </div>
+              ))}
             </div>
-          ))}
-        </div>
-      </OdsFormField>
-      <OdsFormField>
-        <label htmlFor="language" slot="label">
-          {t('zimbra_mailinglist_add_language_label')} *
-        </label>
-        <OdsSelect
-          name="language"
-          value={form.language.value}
-          className="w-1/2"
-          hasError={form.language.hasError}
-          isRequired={form.language.required}
-          data-testid="select-language"
-          onOdsChange={(e) => setValue(e.detail.name, e.detail.value)}
-          defaultValue=""
-          placeholder={t('zimbra_mailinglist_add_select_language_placeholder')}
-        >
-          {languageList?.map((lang) => (
-            <option key={lang} value={lang}>
-              {lang}
-            </option>
-          ))}
-        </OdsSelect>
-      </OdsFormField>
-      <OdsFormField>
-        <label htmlFor="moderationOption" slot="label">
-          {t('zimbra_mailinglist_add_moderation_choice_label')}
-        </label>
-        <div className="flex flex-col gap-4">
-          {moderationChoices.map(({ value, key }) => (
-            <div key={value} className="flex leading-none gap-4">
-              <OdsRadio
-                id={value}
-                name="moderationOption"
-                value={value}
-                onOdsChange={(event) =>
-                  setValue('moderationOption', event.detail.value)
-                }
-                data-testid={`radio-moderation-option-${value}`}
-              ></OdsRadio>
-              <label htmlFor={value} className="flex flex-col">
-                <OdsText preset={ODS_TEXT_PRESET.paragraph}>{t(key)}</OdsText>
-              </label>
+          </OdsFormField>
+        )}
+      />
+      <Controller
+        control={control}
+        name="language"
+        render={({ field: { name, value, onChange, onBlur } }) => (
+          <OdsFormField error={errors?.[name]?.message}>
+            <label htmlFor={name} slot="label">
+              {t('zimbra_mailinglist_add_language_label')} *
+            </label>
+            <OdsSelect
+              data-testid="select-language"
+              placeholder={t(
+                'zimbra_mailinglist_add_select_language_placeholder',
+              )}
+              className="w-1/2"
+              id={name}
+              name={name}
+              value={value}
+              hasError={!!errors[name]}
+              onOdsChange={onChange}
+              onOdsBlur={onBlur}
+            >
+              {languages?.map((lang) => (
+                <option key={lang} value={lang}>
+                  {lang}
+                </option>
+              ))}
+            </OdsSelect>
+          </OdsFormField>
+        )}
+      />
+      <Controller
+        control={control}
+        name="moderationOption"
+        render={({ field: { name, value, onChange } }) => (
+          <OdsFormField error={errors?.[name]?.message}>
+            <label htmlFor={name} slot="label">
+              {t('zimbra_mailinglist_add_moderation_choice_label')} *
+            </label>
+            <div className="flex flex-col gap-4">
+              {moderationChoices.map((choice) => (
+                <div key={choice.value} className="flex leading-none gap-4">
+                  <OdsRadio
+                    id={choice.value}
+                    name={choice.value}
+                    value={value}
+                    isChecked={value === choice.value}
+                    onClick={() => onChange(choice.value)}
+                    data-testid={`radio-moderation-option-${choice.value}`}
+                  ></OdsRadio>
+                  <label
+                    htmlFor={choice.value}
+                    className="flex flex-col cursor-pointer"
+                  >
+                    <OdsText preset={ODS_TEXT_PRESET.paragraph}>
+                      {t(choice.key)}
+                    </OdsText>
+                  </label>
+                </div>
+              ))}
+              <Controller
+                control={control}
+                name="subscriberModeration"
+                render={({ field }) => (
+                  <div className="flex">
+                    <OdsToggle
+                      className="mt-4"
+                      data-testid="toggle-subscriber-moderation"
+                      onOdsChange={(e) => field.onChange(e.detail.value)}
+                      withLabel
+                      name={field.name}
+                    ></OdsToggle>
+                    <div className="flex flex-col ml-4 w-full">
+                      <OdsText preset={ODS_TEXT_PRESET.paragraph}>
+                        {t('zimbra_mailinglist_add_subscriber_moderation')}
+                        <OdsIcon
+                          className="ml-3 text-xs"
+                          id="subs-tooltip-trigger"
+                          name={ODS_ICON_NAME.circleInfo}
+                        />
+                      </OdsText>
+                      <OdsTooltip triggerId="subs-tooltip-trigger">
+                        <OdsText preset={ODS_TEXT_PRESET.paragraph}>
+                          {t(
+                            'zimbra_mailinglist_add_subscriber_moderation_tooltip',
+                          )}
+                        </OdsText>
+                      </OdsTooltip>
+                      <OdsText preset={ODS_TEXT_PRESET.caption}>
+                        {t(
+                          'zimbra_mailinglist_add_subscriber_moderation_info',
+                          {
+                            max: 250,
+                          },
+                        )}
+                      </OdsText>
+                    </div>
+                  </div>
+                )}
+              />
             </div>
-          ))}
-          <div className="flex">
-            <OdsToggle
-              className="mt-4"
-              data-testid="toggle-subscriber-moderation"
-              onOdsChange={() =>
-                setValue(
-                  'subscriberModeration',
-                  form.subscriberModeration.value === 'true' ? 'false' : 'true',
-                )
-              }
-              value={form.subscriberModeration.value === 'true'}
-              defaultValue={
-                (form.subscriberModeration.value as unknown) as
-                  | (false & readonly string[])
-                  | (true & readonly string[]) // this type is broken is ODS
-              }
-              withLabel
-              name="toggle-subscriber-moderation"
-            ></OdsToggle>
-            <div className="flex flex-col ml-4 w-full">
-              <div>
-                <OdsText preset={ODS_TEXT_PRESET.paragraph}>
-                  {t('zimbra_mailinglist_add_subscriber_moderation')}
-                  <OdsIcon
-                    className="ml-3 text-xs"
-                    id="tooltip-trigger"
-                    name={ODS_ICON_NAME.circleInfo}
-                  />
-                </OdsText>
-                <OdsTooltip triggerId="tooltip-trigger">
-                  <OdsText preset={ODS_TEXT_PRESET.paragraph}>
-                    {t('zimbra_mailinglist_add_subscriber_moderation_tooltip')}
-                  </OdsText>
-                </OdsTooltip>
-              </div>
-              <OdsText preset={ODS_TEXT_PRESET.caption}>
-                {t('zimbra_mailinglist_add_subscriber_moderation_info', {
-                  max: 250,
-                })}
-              </OdsText>
-            </div>
-          </div>
-        </div>
-      </OdsFormField>
+          </OdsFormField>
+        )}
+      />
       <div className="flex space-x-5">
         <OdsButton
           slot="actions"
+          type="submit"
           variant={ODS_BUTTON_VARIANT.default}
-          isDisabled={!isFormValid}
+          isDisabled={!isDirty || !isValid}
           isLoading={isSending}
-          onClick={handleSavelick}
           data-testid="confirm-btn"
           label={t('common:confirm')}
         />
@@ -485,6 +478,6 @@ export default function MailingListSettings({
           />
         )}
       </div>
-    </div>
+    </form>
   );
 }